Relocate getRouteKeys() and containsUnsupportedRouteTags() (#22179)
* Fix fitness_trail in OsmRouteType

* Relocate getRouteKeys() and containsUnsupportedRouteTags()
